// frontend/src/services/generated/model/userEntity.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Shanda Automotive API
 * Marketplace connecting vehicle owners with service providers
 * OpenAPI spec version: 1.0
 */
import type { UserEntityProviderStatus } from './userEntityProviderStatus';

export interface UserEntity {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  roles: string[];
  onboardingComplete: boolean;
  providerOnboardingComplete: boolean;
  providerStatus?: UserEntityProviderStatus;
  /** @nullable */
  stripeAccountId?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zipCode?: string | null;
  /** @nullable */
  businessName?: string | null;
  serviceTypes: string[];
  /** @nullable */
  yearsInBusiness?: number | null;
  /** @nullable */
  shopAddress?: string | null;
  /** @nullable */
  shopCity?: string | null;
  /** @nullable */
  shopState?: string | null;
  /** @nullable */
  shopZipCode?: string | null;
  serviceArea?: string[];
  isMobileService: boolean;
  isShopService: boolean;
  shopPhotos: string[];
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  createdAt: string;
  updatedAt: string;
  password: string;
}